aniado validacion formularios php

+++ backend/controller/CategoryController.php
@@ -1,6 +1,13 @@
 <?php
+
+/**
+ *
+ */
 class CategoryController
 {
+    /**
+     * @return array|null
+     */
     public function categorias()
     {
         if (isset($_POST["cantidad"])) {
@@ -11,6 +18,10 @@ class CategoryController
         $dbObjecto = new bdObjeto();
         return $dbObjecto->getCategorias($cantidad);
     }
+
+    /**
+     * @return false|string|null
+     */
     public function categoriasJSON()
     {
         $dbObjeto = new bdObjeto();
@@ -21,16 +32,28 @@ class CategoryController
         }
         return $dbObjeto->getCategoriasJSON($cantidad);
     }
+
+    /**
+     * @return false|string|null
+     */
     public function categorias2()
     {
         $dbObjecto = new bdObjeto();
         return $dbObjecto->getCategorias2();
     }
+
+    /**
+     * @return int|mixed|null
+     */
     public function idCat()
     {
         $db = new bdCategoria();
         return $db->getMaxId();
     }
+
+    /**
+     * @return false|string|null
+     */
     public function categories()
     {
         $dbCategoria = new bdCategoria();
@@ -46,6 +69,11 @@ class CategoryController
         }
         return $dbCategoria->read($campo, $principio, 10);
     }
+
+    /**
+     * @param $id
+     * @return void
+     */
     public function eliminarCategoria($id)
     {
         $dbCategoria = new bdCategoria();
@@ -53,6 +81,10 @@ class CategoryController
             header("Location: ../admin/categories");
         }
     }
+
+    /**
+     * @return array
+     */
     public function getDatosTablaCat()
     {
         $dbUser = new bdCategoria();
@@ -60,6 +92,11 @@ class CategoryController
         $info = [$cabecera2];
         return $info;
     }
+
+    /**
+     * @param $info
+     * @return void
+     */
     public function showEditCat($info)
     {
         $dbCategoria = new bdCategoria();
@@ -67,34 +104,59 @@ class CategoryController
         $info = array_values(json_decode(json_encode($cat), true));
         require "view/editarCategoria.php";
     }
+
+    /**
+     * @return void
+     */
     public function aniadirCat()
     {
         $dbCategoria = new bdCategoria();
         $info = $dbCategoria->getMaxIdCat() + 1;
         require "view/aniadirCategoria.php";
     }
+
+    /**
+     * @return void
+     */
     public function processaniadirCat()
     {
         if ($_SERVER["REQUEST_METHOD"] == "POST") {
             if (isset($_POST["submit"])) {
-                $nombre = $_POST["nombre"];
-                $descripcion = $_POST["descripcion"];
+                if (isset($_POST["nombre"]) && !empty($_POST["nombre"]) && preg_match("/^[A-ZÁÉÍÓÚ][a-záéíóú]+$/", $_POST["nombre"])) {
+                    $nombre = $_POST["nombre"];
+                } else {
+                    header("Location: ../admin/products/aniadirCategories");
+                }
+                if (isset($_POST["descripcion"]) && !empty($_POST["descripcion"]) && preg_match("/^[a-zA-Z0-9]+(?:[\s.]+[a-zA-Z0-9]+)*$/g", $_POST["descripcion"])) {
+                    $descripcion = $_POST["descripcion"];
+                } else {
+                    header("Location: ../admin/products/aniadirCategories");
+                }
                 $dbCategoria = new bdCategoria();
                 $id = $dbCategoria->getMaxIdCat() + 1;
                 mkdir($_SERVER['DOCUMENT_ROOT'] . "/web/backend/view/imgCategories/" . $id);
-                $temporal = $_FILES["imagen"]["tmp_name"];
-                $fileName = $_FILES["imagen"]["name"];
-                $path = $_SERVER['DOCUMENT_ROOT'] . "/web/backend/view/imgCategories/" . $id . "/{$fileName}";
-                $absolutePath = __DIR__ . '/' . $path;
-                if (!file_exists($absolutePath)) {
-                    move_uploaded_file($temporal, $path);
-                    $imagen = "http://localhost/web/backend/view/imgCategories/" . $id . "/{$fileName}";
+                if (isset($_FILES["imagen"]["name"]) && !empty($_FILES["imagen"]["name"])) {
+                    $temporal = $_FILES["imagen"]["tmp_name"];
+                    $fileName = $_FILES["imagen"]["name"];
+                    $path = $_SERVER['DOCUMENT_ROOT'] . "/web/backend/view/imgCategories/" . $id . "/{$fileName}";
+                    $absolutePath = __DIR__ . '/' . $path;
+                    if (!file_exists($absolutePath)) {
+                        move_uploaded_file($temporal, $path);
+                        $imagen = "http://localhost/web/backend/view/imgCategories/" . $id . "/{$fileName}";
+                    }
+                } else {
+                    $imagen = NULL;
                 }
+
                 $dbCategoria->create($nombre, $descripcion, $imagen);
                 header("Location: ../admin/categories");
             }
         }
     }
+
+    /**
+     * @return void
+     */
     public function processCategoria()
     {
         if ($_SERVER['REQUEST_METHOD'] == 'POST') {
@@ -105,25 +167,39 @@ class CategoryController
                     header("Location: ../../admin/categories");
                 }
                 $dbCategoria = new bdCategoria();
-                $temporal = $_FILES["imagen"]["tmp_name"];
-                $fileName = $_FILES["imagen"]["name"];
-                $path = $_SERVER['DOCUMENT_ROOT'] . "/web/backend/view/imgCategories/" . $_POST['idCat'] . "/{$fileName}";
+                if (isset($_FILES["imagen"]["name"]) && !empty($_FILES["imagen"]["name"])) {
+                    $temporal = $_FILES["imagen"]["tmp_name"];
+                    $fileName = $_FILES["imagen"]["name"];
+                    $path = $_SERVER['DOCUMENT_ROOT'] . "/web/backend/view/imgCategories/" . $_POST['idCat'] . "/{$fileName}";
 
-                $absolutePath = __DIR__ . '/' . $path;
-                $dir = $_SERVER['DOCUMENT_ROOT'] . "/web/backend/view/imgCategories/" . $_POST['idCat'];
-                if(!is_dir($dir)) {
-                    mkdir($dir);
-                }
-                if (!file_exists($absolutePath)) {
-                    $files = glob($dir . "/*"); //obtenemos todos los nombres de los ficheros
-                    foreach ($files as $file) {
-                        if (is_file($file))
-                            unlink($file); //elimino el fichero
+                    $absolutePath = __DIR__ . '/' . $path;
+                    $dir = $_SERVER['DOCUMENT_ROOT'] . "/web/backend/view/imgCategories/" . $_POST['idCat'];
+                    if (!is_dir($dir)) {
+                        mkdir($dir);
                     }
-                    move_uploaded_file($temporal, $path);
-                    $categoryImg = "http://localhost/web/backend/view/imgCategories/" . $_POST['idCat'] . "/{$fileName}";
+                    if (!file_exists($absolutePath)) {
+                        $files = glob($dir . "/*"); //obtenemos todos los nombres de los ficheros
+                        foreach ($files as $file) {
+                            if (is_file($file))
+                                unlink($file); //elimino el fichero
+                        }
+                        move_uploaded_file($temporal, $path);
+                        $categoryImg = "http://localhost/web/backend/view/imgCategories/" . $_POST['idCat'] . "/{$fileName}";
+                    }
+                } else {
+                    $categoryImg = NULL;
                 }
-                $array = array("Nombre" => $_POST["nombre"], "Descripcion" => $_POST["descripcion"], "Imagen" => $categoryImg,);
+                if (isset($_POST["nombre"]) && !empty($_POST["nombre"]) && preg_match("/^[A-ZÁÉÍÓÚ][a-záéíóú]+$/", $_POST["nombre"])) {
+                    $nombre = $_POST["nombre"];
+                } else {
+                    header("Location: ../admin/products/aniadirCategories");
+                }
+                if (isset($_POST["descripcion"]) && !empty($_POST["descripcion"]) && preg_match("/^[a-zA-Z0-9]+(?:[\s.]+[a-zA-Z0-9]+)*$/g", $_POST["descripcion"])) {
+                    $descripcion = $_POST["descripcion"];
+                } else {
+                    header("Location: ../admin/products/aniadirCategories");
+                }
+                $array = array("Nombre" => $nombre, "Descripcion" => $descripcion, "Imagen" => $categoryImg,);
                 $dbCategoria->update($id, $array);
                 header("Location: ../../admin/categories");
             }
